+++ stories/pricing/cc-pricing-product-consumption.stories.js
@@ -180,6 +180,27 @@ const basePulsar = {
   ],
 };
 
+const baseHeptapod = {
+  name: 'Heptapod',
+  icon: 'https://static-assets.cellar.services.clever-cloud.com/logos/heptapod.svg',
+  sections: [
+    {
+      type: 'storage',
+      intervals: [
+        {
+          minRange: 0,
+          maxRange: 1 * 1e9,
+          price: 0 * THIRTY_DAYS_IN_HOURS / ONE_GIGABYTE,
+        },
+        {
+          minRange: 1 * 1e9,
+          price: 0.00002777777777777778 * THIRTY_DAYS_IN_HOURS / ONE_GIGABYTE,
+        },
+      ],
+    },
+  ],
+};
+
 export const defaultStory = makeStory(conf, {
   items: [baseCellar],
 });
@@ -243,6 +264,10 @@ export const dataLoadedWithFsBucket = makeStory(conf, {
 
 export const dataLoadedWithPulsar = makeStory(conf, {
   items: [basePulsar],
+});
+
+export const dataLoadedWithHeptapod = makeStory(conf, {
+  items: [baseHeptapod],
 });
 
 export const dataLoadedWithCustomHead = makeStory(conf, {
@@ -372,6 +397,7 @@ enhanceStoriesNames({
   errorWithPulsar,
   dataLoadedWithFsBucket,
   dataLoadedWithPulsar,
+  dataLoadedWithHeptapod,
   dataLoadedWithCustomHead,
   dataLoadedWithEmptyHead,
   dataLoadedWithCustomDescription,